move gradient banner generation to pkg/cli
