+++ benchmark/cmd/command/banner.go
@@ -1,11 +1,7 @@
-// Copied from github.com/pyroscope-io/pyroscope/cmd/command/banner.go
 package command
 
 import (
-	"strings"
-
-	"github.com/aybabtme/rgbterm"
-	"github.com/fatih/color"
+	"github.com/pyroscope-io/pyroscope/pkg/cli"
 )
 
 // made here http://patorjk.com/software/taag/#p=display&f=Doom&t=Pyrobench
@@ -25,34 +21,6 @@ func init() {
 	banner = banner[1 : len(banner)-2]
 }
 
-const (
-	startColor = 0xffd651
-	endColor   = 0xf64d3d
-)
-
-func gradient(start, end, offset int, progress float64) uint8 {
-	start = (start >> offset) & 0xff
-	end = (end >> offset) & 0xff
-	return uint8(start + int(float64(end-start)*progress))
-}
-
 func gradientBanner() string {
-	if color.NoColor {
-		return banner + "\n"
-	}
-
-	str := ""
-	arr := strings.Split(banner, "\n")
-	l := len(arr)
-	for i, line := range arr {
-		if line == "" {
-			break
-		}
-		progress := float64(i) / float64(l-1)
-		r := gradient(startColor, endColor, 16, progress)
-		g := gradient(startColor, endColor, 8, progress)
-		b := gradient(startColor, endColor, 0, progress)
-		str += rgbterm.FgString(line, r, g, b) + "\n"
-	}
-	return str
+	return cli.GradientBanner(banner)
 }
